HW-3 Coin UI MVP 패턴 적용
